Add files via upload
Simple script to help learn Mandarin. Designed for people who don't have anyone to practice speech. It should through an error message if you aren't understood at all, and show what you are saying in 2 forms of pinyin and simplified Chinese (if you want traditional Chinese, replace all the zh-CN with zh-TW but not sure if it will work with the pinyin converter).